Move the in-memory canvas to the shared layer
