+++ src/lib/entities/drawing/index.ts
@@ -1,0 +1,2 @@
+export { styleWithBrush } from './model/style-with-brush';
+export { extractToSeparateCanvas } from './model/extract-to-separate-canvas';
