+++ src/lib/entities/drawing/model/types.d.ts
