--- src/lib/features/in-memory-canvas/index.ts
@@ -1,2 +1,0 @@
-export { extractToSeparateCanvas } from './model/extract-last-drawing';
-export { discardInMemoryCanvas } from './model/in-memory-canvas';
