+++ src/lib/shared/in-memory-canvas/index.ts
@@ -1,0 +1,1 @@
+export { requestCanvasToFit, discardInMemoryCanvas } from './model/in-memory-canvas';
